Minerva/requirements.txt
numpy==1.20.3
pubchempy==1.0.4
requests==2.27.1
pyautogui==0.9.53
pywin32
pyserial
scipy
matplotlib==3.7.1
pythonnet==3.0.2
